feat(dev): 改造 TUI 的实现，我希望能参考 gitui 的实现，用更规范的方式来做。请先分析给出产品方案和技术方案后再实施
